+++ src/main/java/com/example/AirbnbBookingSpring/dtos/AirbnbResponseDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class AirbnbResponseDTO {
     private Long id;
     private String name;
     private String description;
-    private double pricePerNight;
+    private BigDecimal pricePerNight;
     private String cityName;
 
 }
